rename check.rs -> comparison.rs
