+++ src/lib.rs
@@ -1,4 +1,4 @@
-pub mod check;
+pub mod comparison;
 pub mod constant;
 pub mod die;
 pub mod operation;
